requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
coverage==7.5.3
cryptography==42.0.7
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-DotEnv==0.1.2
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.3
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
reportlab==4.2.0
rich==13.7.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
tzdata==2024.1
waitress==3.0.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xlrd==2.0.1